Play map & client refactoring.
Play Map:

- Make hovered territory text say Unowned for UNOWNED country sprite
  state (instead of just not showing anything).

- Make country color text update without having to move mouse when the
  country color is changed without using the mouse to click on a country.

- Move play map background image inside play map actor, and get rid of
  the play map stack in ClassicPlayScreen.

- Move territory name text actor inside play map actor.

Client-level refactoring:

- Remove CoordinateSpaces class.

- Remove all uses of fg-tools::Geometry class, and all uses of Geometry
  package classes like Point2D, Scaling2D, etc.

- Replace all of the above with re-usable instances of LibGDX's Vector2
  class, which eliminates garbage collection and provides much better
  functionality.

- The CoordinateSpaces utility class that converts between various
  coordinate spaces was inconsistent and confusing. It's easier to chain
  together several mutations on a Vector2 in the class that needs it.
  Also, the goal is to offload more of the coordinate space calculations
  onto LibGDX via scene2d.ui's Actor class coordinate space conversion
  methods, which are pretty smart as they take into account table layout
  parameters, such as padding, spacing, alignment, etc, which can be
  pretty complicated, fragile, and buggy to compute manually (which I
  learned the hard way - hence this refactoring).

+++ client/src/main/java/com/forerunnergames/peril/client/application/ClientApplicationProperties.java
@@ -109,8 +109,8 @@ public final class ClientApplicationProperties
       }
     }
 
-    GraphicsSettings.INITIAL_WINDOW_WIDTH = parseInteger (WINDOW_WIDTH_PROPERTY_KEY, GraphicsSettings.MIN_INITIAL_WINDOW_WIDTH);
-    GraphicsSettings.INITIAL_WINDOW_HEIGHT = parseInteger (WINDOW_HEIGHT_PROPERTY_KEY, GraphicsSettings.MIN_INITIAL_WINDOW_HEIGHT);
+    GraphicsSettings.INITIAL_WINDOW_WIDTH = parseInteger (WINDOW_WIDTH_PROPERTY_KEY, (int) GraphicsSettings.MIN_INITIAL_WINDOW_WIDTH);
+    GraphicsSettings.INITIAL_WINDOW_HEIGHT = parseInteger (WINDOW_HEIGHT_PROPERTY_KEY, (int) GraphicsSettings.MIN_INITIAL_WINDOW_HEIGHT);
     GraphicsSettings.IS_WINDOW_RESIZABLE = parseBoolean (WINDOW_RESIZABLE_PROPERTY_KEY);
     GraphicsSettings.WINDOW_TITLE = properties.getProperty (WINDOW_TITLE_PROPERTY_KEY);
     GraphicsSettings.IS_FULLSCREEN = parseBoolean (FULLSCREEN_PROPERTY_KEY);
